move models to subfolder
