+++ src/teehr/calculated_fields/__init__.py
@@ -1,7 +1,7 @@
 """Calculated fields package."""
 
-from teehr.calculated_fields.row_level_models import RowLevelCalculatedFields
-from teehr.calculated_fields.timeseries_aware_models import TimeseriesAwareCalculatedFields
+from teehr.calculated_fields.models.row_level import RowLevelCalculatedFields
+from teehr.calculated_fields.models.timeseries_aware import TimeseriesAwareCalculatedFields
 
 __all__ = [
     "RowLevelCalculatedFields",
