+++ src/teehr/calculated_fields/row_level_spark.py
@@ -1,20 +1,13 @@
-"""Row-level calculated-field executors.
-
-This module hosts execution logic for row-level calculated fields so model
-classes can remain declarative containers.
-"""
+"""Spark-native row-level calculated-field executors."""
 
 from __future__ import annotations
 
 from datetime import timedelta
 from typing import Union
 
-import numpy as np
 import pandas as pd
 import pyspark.sql as ps
 import pyspark.sql.functions as F
-import pyspark.sql.types as T
-from pyspark.sql.functions import pandas_udf
 
 
 def apply_month(sdf: ps.DataFrame, *, input_field_name: str, output_field_name: str) -> ps.DataFrame:
@@ -22,39 +15,9 @@ def apply_month(sdf: ps.DataFrame, *, input_field_name: str, output_field_name: 
     return sdf.withColumn(output_field_name, F.month(F.col(input_field_name)))
 
 
-def apply_month_pandas(
-    sdf: ps.DataFrame,
-    *,
-    input_field_name: str,
-    output_field_name: str,
-) -> ps.DataFrame:
-    """Add month number via pandas execution (opt-in path)."""
-
-    @pandas_udf(returnType=T.IntegerType())
-    def _month(ts: pd.Series) -> pd.Series:
-        return pd.to_datetime(ts).dt.month
-
-    return sdf.withColumn(output_field_name, _month(input_field_name))
-
-
 def apply_year(sdf: ps.DataFrame, *, input_field_name: str, output_field_name: str) -> ps.DataFrame:
     """Add year from a timestamp column."""
     return sdf.withColumn(output_field_name, F.year(F.col(input_field_name)))
-
-
-def apply_year_pandas(
-    sdf: ps.DataFrame,
-    *,
-    input_field_name: str,
-    output_field_name: str,
-) -> ps.DataFrame:
-    """Add year via pandas execution (opt-in path)."""
-
-    @pandas_udf(returnType=T.IntegerType())
-    def _year(ts: pd.Series) -> pd.Series:
-        return pd.to_datetime(ts).dt.year
-
-    return sdf.withColumn(output_field_name, _year(input_field_name))
 
 
 def apply_water_year(sdf: ps.DataFrame, *, input_field_name: str, output_field_name: str) -> ps.DataFrame:
@@ -66,22 +29,6 @@ def apply_water_year(sdf: ps.DataFrame, *, input_field_name: str, output_field_n
             + F.when(F.month(F.col(input_field_name)) >= 10, F.lit(1)).otherwise(F.lit(0))
         ).cast("int"),
     )
-
-
-def apply_water_year_pandas(
-    sdf: ps.DataFrame,
-    *,
-    input_field_name: str,
-    output_field_name: str,
-) -> ps.DataFrame:
-    """Add water year via pandas execution (opt-in path)."""
-
-    @pandas_udf(returnType=T.IntegerType())
-    def _water_year(ts: pd.Series) -> pd.Series:
-        dt = pd.to_datetime(ts)
-        return (dt.dt.year + (dt.dt.month >= 10).astype(int)).astype("int32")
-
-    return sdf.withColumn(output_field_name, _water_year(input_field_name))
 
 
 def apply_normalized_flow(
@@ -98,28 +45,6 @@ def apply_normalized_flow(
             F.col(primary_value_field_name).cast("double"),
             F.col(drainage_area_field_name).cast("double"),
         ).cast("float"),
-    )
-
-
-def apply_normalized_flow_pandas(
-    sdf: ps.DataFrame,
-    *,
-    primary_value_field_name: str,
-    drainage_area_field_name: str,
-    output_field_name: str,
-) -> ps.DataFrame:
-    """Normalize flow by drainage area via pandas execution (opt-in path)."""
-
-    @pandas_udf(returnType=T.FloatType())
-    def _norm(v: pd.Series, a: pd.Series) -> pd.Series:
-        v = pd.to_numeric(v, errors="coerce")
-        a = pd.to_numeric(a, errors="coerce")
-        out = np.divide(v, a, out=np.full(len(v), np.nan, dtype="float64"), where=(a != 0))
-        return pd.Series(out, index=v.index).astype("float32")
-
-    return sdf.withColumn(
-        output_field_name,
-        _norm(primary_value_field_name, drainage_area_field_name),
     )
 
 
@@ -141,27 +66,6 @@ def apply_seasons(
     else:
         expr = expr.otherwise(F.lit(None).cast("string"))
     return sdf.withColumn(output_field_name, expr)
-
-
-def apply_seasons_pandas(
-    sdf: ps.DataFrame,
-    *,
-    value_time_field_name: str,
-    season_months: dict,
-    output_field_name: str,
-) -> ps.DataFrame:
-    """Map month to season label via pandas execution (opt-in path)."""
-    month_to_season = {}
-    for season, months in season_months.items():
-        for month in months:
-            month_to_season[int(month)] = season
-
-    @pandas_udf(returnType=T.StringType())
-    def _seasons(ts: pd.Series) -> pd.Series:
-        dt = pd.to_datetime(ts)
-        return dt.dt.month.map(month_to_season)
-
-    return sdf.withColumn(output_field_name, _seasons(value_time_field_name))
 
 
 def apply_forecast_lead_time(
@@ -263,6 +167,9 @@ def apply_forecast_lead_time_bins(
     bin_size: Union[pd.Timedelta, timedelta, str, list, dict],
 ) -> ps.DataFrame:
     """Add forecast lead-time bin IDs with configurable binning."""
+    import pyspark.sql.types as T
+    from pyspark.sql.functions import pandas_udf
+
     normalized_bin_size = validate_forecast_lead_time_bin_size(bin_size)
 
     if lead_time_field_name not in sdf.columns:
@@ -347,24 +254,6 @@ def apply_threshold_value_exceeded(
     )
 
 
-def apply_threshold_value_exceeded_pandas(
-    sdf: ps.DataFrame,
-    *,
-    input_field_name: str,
-    threshold_field_name: str,
-    output_field_name: str,
-) -> ps.DataFrame:
-    """Add value > threshold via pandas execution (opt-in path)."""
-
-    @pandas_udf(returnType=T.BooleanType())
-    def _thr(v: pd.Series, t: pd.Series) -> pd.Series:
-        v = pd.to_numeric(v, errors="coerce")
-        t = pd.to_numeric(t, errors="coerce")
-        return (v > t).fillna(False)
-
-    return sdf.withColumn(output_field_name, _thr(input_field_name, threshold_field_name))
-
-
 def apply_threshold_value_not_exceeded(
     sdf: ps.DataFrame,
     *,
@@ -380,24 +269,6 @@ def apply_threshold_value_not_exceeded(
             F.lit(False),
         ).cast("boolean"),
     )
-
-
-def apply_threshold_value_not_exceeded_pandas(
-    sdf: ps.DataFrame,
-    *,
-    input_field_name: str,
-    threshold_field_name: str,
-    output_field_name: str,
-) -> ps.DataFrame:
-    """Add value <= threshold via pandas execution (opt-in path)."""
-
-    @pandas_udf(returnType=T.BooleanType())
-    def _thr(v: pd.Series, t: pd.Series) -> pd.Series:
-        v = pd.to_numeric(v, errors="coerce")
-        t = pd.to_numeric(t, errors="coerce")
-        return (v <= t).fillna(False)
-
-    return sdf.withColumn(output_field_name, _thr(input_field_name, threshold_field_name))
 
 
 def apply_day_of_year(sdf: ps.DataFrame, *, input_field_name: str, output_field_name: str) -> ps.DataFrame:
@@ -417,32 +288,6 @@ def apply_day_of_year(sdf: ps.DataFrame, *, input_field_name: str, output_field_
     return sdf.withColumn(output_field_name, adjusted.cast("int"))
 
 
-def apply_day_of_year_pandas(
-    sdf: ps.DataFrame,
-    *,
-    input_field_name: str,
-    output_field_name: str,
-) -> ps.DataFrame:
-    """Add leap-year-adjusted day-of-year via pandas execution (opt-in path)."""
-
-    @pandas_udf(returnType=T.IntegerType())
-    def _doy(ts: pd.Series) -> pd.Series:
-        dt = pd.to_datetime(ts)
-        year = dt.dt.year
-        month = dt.dt.month
-        day = dt.dt.day
-        doy = dt.dt.dayofyear
-        is_leap = ((year % 4 == 0) & ((year % 100 != 0) | (year % 400 == 0)))
-        out = np.where(
-            is_leap & (month == 2) & (day == 29),
-            59,
-            np.where(is_leap & (month > 2), doy - 1, doy),
-        )
-        return pd.Series(out, index=ts.index).astype("int32")
-
-    return sdf.withColumn(output_field_name, _doy(input_field_name))
-
-
 def apply_hour_of_year(sdf: ps.DataFrame, *, input_field_name: str, output_field_name: str) -> ps.DataFrame:
     """Add leap-year-adjusted hour-of-year index."""
     dt_col = F.col(input_field_name)
@@ -459,33 +304,6 @@ def apply_hour_of_year(sdf: ps.DataFrame, *, input_field_name: str, output_field
         .otherwise((doy_col - F.lit(1)) * F.lit(24) + hour_col)
     )
     return sdf.withColumn(output_field_name, adjusted.cast("int"))
-
-
-def apply_hour_of_year_pandas(
-    sdf: ps.DataFrame,
-    *,
-    input_field_name: str,
-    output_field_name: str,
-) -> ps.DataFrame:
-    """Add leap-year-adjusted hour-of-year via pandas execution (opt-in path)."""
-
-    @pandas_udf(returnType=T.IntegerType())
-    def _hoy(ts: pd.Series) -> pd.Series:
-        dt = pd.to_datetime(ts)
-        year = dt.dt.year
-        month = dt.dt.month
-        day = dt.dt.day
-        doy = dt.dt.dayofyear
-        hour = dt.dt.hour
-        is_leap = ((year % 4 == 0) & ((year % 100 != 0) | (year % 400 == 0)))
-        out = np.where(
-            is_leap & (month == 2) & (day == 29),
-            58 * 24 + hour,
-            np.where(is_leap & (month > 2), (doy - 2) * 24 + hour, (doy - 1) * 24 + hour),
-        )
-        return pd.Series(out, index=ts.index).astype("int32")
-
-    return sdf.withColumn(output_field_name, _hoy(input_field_name))
 
 
 def apply_generic_sql(sdf: ps.DataFrame, *, output_field_name: str, sql_statement: str) -> ps.DataFrame:
