+++ src/teehr/metrics/metric_attributes.py
@@ -1,5 +1,5 @@
 """Dictionaries defining static metric attributes."""
-from teehr.metrics.base_models import MetricCategories as mc
+from teehr.metrics.models.base import MetricCategories as mc
 
 
 ME_ATTRS = {
