+++ src/teehr/metrics/models/bootstrap.py
@@ -6,7 +6,7 @@ from numpy.random import RandomState
 from pydantic import Field
 
 import teehr.metrics.bootstrap_funcs as bootstrap_funcs
-from teehr.metrics.base_models import BootstrapBasemodel
+from teehr.metrics.models.base import BootstrapBasemodel
 
 
 class Gumboot(BootstrapBasemodel):
