+++ src/teehr/metrics/models/probabilistic.py
@@ -3,10 +3,10 @@ from typing import Callable, Dict
 
 from pydantic import Field
 
-from teehr.metrics.base_models import ProbabilisticBasemodel
+from teehr.metrics.models.base import ProbabilisticBasemodel
 from teehr.metrics import probabilistic_funcs
 import teehr.metrics.metric_attributes as tma
-from teehr.metrics.base_models import CRPSEstimators
+from teehr.metrics.models.base import CRPSEstimators
 
 
 class CRPS(ProbabilisticBasemodel):
