+++ src/teehr/metrics/signature_funcs.py
@@ -2,8 +2,8 @@
 import pandas as pd
 import numpy as np
 
-from teehr.metrics.base_models import MetricsBasemodel
-from teehr.metrics.base_models import TransformEnum
+from teehr.metrics.models.base import MetricsBasemodel
+from teehr.metrics.models.base import TransformEnum
 from typing import Callable, Optional
 import logging
 logger = logging.getLogger(__name__)
