+++ tests/data/test_warehouse_setup/create_large_ensemble_test_warehouse.py
@@ -8,7 +8,7 @@ from shapely.geometry import Point
 import geopandas as gpd
 
 import teehr
-import teehr.calculated_fields.row_level_models as rcf
+import teehr.calculated_fields.models.row_level as rcf
 
 
 def _create_warehouse(dir_path):
