+++ tests/data/test_warehouse_setup/create_three_location_test_warehouse.py
@@ -4,7 +4,7 @@ import tempfile
 from pathlib import Path
 
 import teehr
-import teehr.calculated_fields.row_level_models as rcf
+import teehr.calculated_fields.models.row_level as rcf
 
 
 def _create_warehouse(dir_path):
